Made adding two numbers possible and made NumExp contain doubles

+++ src/AddExp.java
@@ -16,7 +16,7 @@ public class AddExp extends Exp{
         System.out.print(")");
     }
 
-    public int eval(){
+    public double eval(){
         return left.eval() + right.eval();
     }
